+++ ChangeLog.md
@@ -1,6 +1,7 @@
 # ChangeLog for EconomyDiffEq
 
 0.0.4 add the country name etc. to the database 
+    all tables have (Maybe Double) to indicate where observations are missing
 
 
 0.0.3 separate folder for each study. (baseTest, pt)